refactor(schemas): migrate Phase 5 endpoint schemas to domain schema files (#6042)
Move locally-defined Pydantic models from 6 endpoint files into domain schema files:
- schemas_workflows.py: ApprovalGate schemas (approval_gates.py)
- schemas_code.py: OAI/ChatCompletion schemas (openai_compat.py)
- schemas_analytics.py: RUM metric schemas (rum.py)
- schemas_agent.py: AgentOrg + Collaboration schemas (agent_org.py, collaboration.py)
- schemas_knowledge.py: ConversationFile schemas (conversation_files.py)
