+++ autobot-backend/api/agent_org.py
@@ -12,76 +12,25 @@ import logging
 from typing import Any, Dict, List, Optional
 
 from fastapi import APIRouter, Depends, HTTPException, Query, status
-from pydantic import BaseModel, Field
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from api.schemas_agent import (
+    AgentDelegateRequest,
+    AgentSummary,
+    ChainOfCommandResponse,
+    DelegationResponse,
+    DelegationStatusUpdate,
+    OrgNodeResponse,
+    UpdateOrgRequest,
+    UpsertOrgRequest,
+)
 from api.user_management.dependencies import get_db_session
+from autobot_shared.error_boundaries import ErrorCategory, with_error_handling
 from services.agent_org_service import AgentOrgService
 from services.delegation_service import DelegationService
-from autobot_shared.error_boundaries import ErrorCategory, with_error_handling
 
 logger = logging.getLogger(__name__)
 router = APIRouter()
-
-
-# -- Schemas ---------------------------------------------------------------
-
-
-class OrgNodeResponse(BaseModel):
-    """Single node in the org tree response (#1405)."""
-
-    agent_id: str
-    name: str
-    org_role: str
-    title: Optional[str] = None
-    capabilities: Optional[str] = None
-    direct_reports_count: int = 0
-    children: List["OrgNodeResponse"] = Field(default_factory=list)
-
-
-OrgNodeResponse.model_rebuild()
-
-
-class AgentSummary(BaseModel):
-    """Compact agent summary used in chain of command (#1405)."""
-
-    agent_id: str
-    name: str
-    org_role: str
-    title: Optional[str] = None
-
-
-class ChainOfCommandResponse(BaseModel):
-    """Ordered list from agent to org root (#1405)."""
-
-    chain: List[AgentSummary]
-
-
-class UpdateOrgRequest(BaseModel):
-    """Request body for PATCH /agents/{agent_id}/org (#1405)."""
-
-    reports_to: Optional[str] = Field(
-        default=None,
-        description="agent_id of the new manager, or null to clear",
-    )
-    org_role: Optional[str] = Field(
-        default=None,
-        description="One of: manager, coordinator, specialist, worker",
-    )
-    title: Optional[str] = Field(default=None, description="Human-readable job title")
-    capabilities: Optional[str] = Field(
-        default=None, description="Free-text capability description"
-    )
-
-
-class UpsertOrgRequest(BaseModel):
-    """Request body for PUT /agents/{agent_id}/org (#1405)."""
-
-    name: str
-    org_role: str = "worker"
-    reports_to: Optional[str] = None
-    title: Optional[str] = None
-    capabilities: Optional[str] = None
 
 
 # -- Endpoints -------------------------------------------------------------
@@ -261,38 +210,6 @@ async def get_role_defaults(
     return svc.get_role_defaults(role)
 
 
-# -- Delegation schemas (#1753) -------------------------------------------
-
-
-class DelegateRequest(BaseModel):
-    """Request body for POST /{manager_id}/delegate (#1753)."""
-
-    assignee_id: str = Field(..., description="Direct report to assign to")
-    task_description: str = Field(..., description="What the assignee should do")
-    context: Optional[Dict[str, Any]] = Field(
-        default=None, description="Extra context for the task"
-    )
-
-
-class DelegationResponse(BaseModel):
-    """Response for a task delegation (#1753)."""
-
-    id: str
-    delegator_id: str
-    assignee_id: str
-    task_description: str
-    status: str
-    escalated_to: Optional[str] = None
-    created_at: Optional[str] = None
-
-
-class DelegationStatusUpdate(BaseModel):
-    """Request body for PATCH /delegations/{id}/status (#1753)."""
-
-    status: str = Field(..., description="New status value")
-    result: Optional[Dict[str, Any]] = None
-
-
 def _delegation_to_response(d) -> DelegationResponse:
     """Convert TaskDelegation ORM row to response (#1753)."""
     return DelegationResponse(
@@ -322,7 +239,7 @@ def _delegation_to_response(d) -> DelegationResponse:
 )
 async def delegate_task(
     manager_id: str,
-    body: DelegateRequest,
+    body: AgentDelegateRequest,
     session: AsyncSession = Depends(get_db_session),
 ) -> DelegationResponse:
     """
