+++ autobot-backend/api/rum.py
@@ -14,8 +14,19 @@ from datetime import datetime, timedelta, timezone
 from typing import List, Optional
 
 from fastapi import APIRouter, HTTPException
-from pydantic import BaseModel
-
+from api.schemas_analytics import (
+    RumApiCallMetric,
+    RumConfig,
+    RumCriticalIssueMetric,
+    RumEvent,
+    RumJsErrorMetric,
+    RumMetrics,
+    RumPageMetrics,
+    RumResourceMetric,
+    RumSessionMetric,
+    RumUserActionMetric,
+    RumWebSocketMetric,
+)
 from api.schemas_workflows import (
     RUMClearResponse,
     RUMConfigResponse,
@@ -136,117 +147,6 @@ def _log_rum_event_by_type(
         return
 
     rum_logger.info(log_message)
-
-
-class RumEvent(BaseModel):
-    type: str
-    timestamp: str
-    sessionId: str
-    url: str
-    userAgent: str
-    data: Metadata = {}
-
-
-class RumConfig(BaseModel):
-    enabled: bool = False
-    error_tracking: bool = True
-    performance_monitoring: bool = True
-    interaction_tracking: bool = False
-    session_recording: bool = False
-    sample_rate: int = 100
-    max_events_per_session: int = 1000
-    debug_mode: bool = False
-    log_to_backend: bool = True
-    log_level: str = "info"
-
-
-# Issue #476: RUM Metrics models for Prometheus integration
-class PageMetrics(BaseModel):
-    """Page performance metrics from frontend."""
-
-    page: str
-    load_time_seconds: Optional[float] = None
-    fcp_seconds: Optional[float] = None  # First Contentful Paint
-    lcp_seconds: Optional[float] = None  # Largest Contentful Paint
-    tti_seconds: Optional[float] = None  # Time to Interactive
-    dom_loaded_seconds: Optional[float] = None
-
-
-class ApiCallMetric(BaseModel):
-    """Frontend API call metric."""
-
-    endpoint: str
-    method: str
-    status: str  # success, error, timeout
-    latency_seconds: float
-    is_slow: bool = False
-    is_timeout: bool = False
-    error_type: Optional[str] = None  # network, http, timeout
-
-
-class JsErrorMetric(BaseModel):
-    """JavaScript error metric."""
-
-    error_type: str  # syntax, reference, type, etc.
-    page: str
-    is_rejection: bool = False  # True for unhandled Promise rejections
-    component: Optional[str] = None  # Vue component name if applicable
-
-
-class UserActionMetric(BaseModel):
-    """User action/interaction metric."""
-
-    action_type: str  # click, submit, navigation, etc.
-    page: str
-    form_name: Optional[str] = None
-    form_status: Optional[str] = None  # success, validation_error, server_error
-
-
-class SessionMetric(BaseModel):
-    """Session metric."""
-
-    event: str  # start, end
-    duration_seconds: Optional[float] = None
-
-
-class WebSocketMetric(BaseModel):
-    """WebSocket event metric from frontend."""
-
-    event: str  # connect, disconnect, error, reconnect
-    direction: Optional[str] = None  # sent, received
-    event_type: Optional[str] = None  # message type
-
-
-class ResourceMetric(BaseModel):
-    """Resource load metric."""
-
-    resource_type: str  # script, stylesheet, image, font
-    load_time_seconds: float
-    is_slow: bool = False
-
-
-class CriticalIssueMetric(BaseModel):
-    """Critical issue from frontend."""
-
-    issue_type: str  # api_timeout, error, crash
-
-
-class RumMetrics(BaseModel):
-    """
-    Batch of RUM metrics from frontend.
-    Issue #476: Used for Prometheus metrics export.
-    """
-
-    session_id: str
-    timestamp: str
-    page_metrics: Optional[PageMetrics] = None
-    api_calls: Optional[List[ApiCallMetric]] = None
-    js_errors: Optional[List[JsErrorMetric]] = None
-    user_actions: Optional[List[UserActionMetric]] = None
-    session: Optional[SessionMetric] = None
-    websocket_events: Optional[List[WebSocketMetric]] = None
-    resources: Optional[List[ResourceMetric]] = None
-    critical_issues: Optional[List[CriticalIssueMetric]] = None
 
 
 def setup_rum_logger():
@@ -573,7 +473,7 @@ def format_rum_log_message(event_data: Metadata) -> str:
 # =============================================================================
 
 
-def _process_page_metrics(metrics_manager, page_metrics: PageMetrics) -> None:
+def _process_page_metrics(metrics_manager, page_metrics: RumPageMetrics) -> None:
     """Process page performance metrics."""
     if page_metrics.load_time_seconds is not None:
         metrics_manager.record_frontend_page_load(
@@ -591,7 +491,7 @@ def _process_page_metrics(metrics_manager, page_metrics: PageMetrics) -> None:
         )
 
 
-def _process_api_calls(metrics_manager, api_calls: List[ApiCallMetric]) -> None:
+def _process_api_calls(metrics_manager, api_calls: List[RumApiCallMetric]) -> None:
     """Process API call metrics."""
     for call in api_calls:
         metrics_manager.record_frontend_api_request(
@@ -608,7 +508,7 @@ def _process_api_calls(metrics_manager, api_calls: List[ApiCallMetric]) -> None:
             )
 
 
-def _process_js_errors(metrics_manager, js_errors: List[JsErrorMetric]) -> None:
+def _process_js_errors(metrics_manager, js_errors: List[RumJsErrorMetric]) -> None:
     """Process JavaScript error metrics."""
     for error in js_errors:
         if error.is_rejection:
@@ -622,7 +522,7 @@ def _process_js_errors(metrics_manager, js_errors: List[JsErrorMetric]) -> None:
 
 
 def _process_user_actions(
-    metrics_manager, user_actions: List[UserActionMetric]
+    metrics_manager, user_actions: List[RumUserActionMetric]
 ) -> None:
     """Process user action metrics."""
     for action in user_actions:
@@ -633,7 +533,7 @@ def _process_user_actions(
             )
 
 
-def _process_session_metric(metrics_manager, session: SessionMetric) -> None:
+def _process_session_metric(metrics_manager, session: RumSessionMetric) -> None:
     """Process session metrics."""
     if session.event == "start":
         metrics_manager.record_frontend_session_start()
@@ -642,7 +542,7 @@ def _process_session_metric(metrics_manager, session: SessionMetric) -> None:
 
 
 def _process_websocket_events(
-    metrics_manager, websocket_events: List[WebSocketMetric]
+    metrics_manager, websocket_events: List[RumWebSocketMetric]
 ) -> None:
     """Process WebSocket event metrics."""
     for ws_event in websocket_events:
@@ -653,7 +553,7 @@ def _process_websocket_events(
             )
 
 
-def _process_resources(metrics_manager, resources: List[ResourceMetric]) -> None:
+def _process_resources(metrics_manager, resources: List[RumResourceMetric]) -> None:
     """Process resource load metrics."""
     for resource in resources:
         metrics_manager.record_frontend_resource_load(
@@ -664,7 +564,7 @@ def _process_resources(metrics_manager, resources: List[ResourceMetric]) -> None
 
 
 def _process_critical_issues(
-    metrics_manager, critical_issues: List[CriticalIssueMetric]
+    metrics_manager, critical_issues: List[RumCriticalIssueMetric]
 ) -> None:
     """Process critical issue metrics."""
     for issue in critical_issues:
